animal_div_server/requirements.txt
fastapi
keras.models
Pillow
numpy